Add android.test.runner.stubs dependency
In preparation for the removal of the non-junit classes in the
android.test.runner library from the android.jar this adds a dependency
on android.test.runner.stubs to ensure this code will continue to
compile.

The following change descriptions were generated automatically and so
may be a little repetitive. They are provided to give the reviewer
enough information to check the comments match what has actually been
changed and check the reasoning behind the changes.

* tests/unit/Android.mk
    Added 'android.test.runner.stubs' to LOCAL_JAVA_LIBRARIES because
    TVUnitTests's will needs them in order to compile once
    its classes are removed from the system_current SDK on which it
    currently depends.

Bug: 30188076
Test: make checkbuild
Change-Id: I6de67f924f16d7450098ecf19c21d5ce2e0c73fd