Fix link-type warnings
Modules under this directory had link-type warnings which will become
errors in the future. Let's fix them.

* Some prebuilts libs (jsr330-jar, guava-22-0-jar) are replaced to the
modules built from the source.

* Rest of the prebuilt libs are added with LOCAL_SDK_VERSION := current
so that they can be linked from apps built with SDK. Otherwise, the
prebuilt libs are recognized as built with platform internal APIs and
thus can't be linked from SDK-based apps.

* The test apks are built with system SDK (LOCAL_SDK_VERSION :=
system_current) because tv-test-common which is commonly used by the
apks are built with system SDK.

* LOCAL_MIN_SDK_VERSION were removed from the test apks. This is because
robolectric that is used by the apps is using MethodHandle.invoke which
can be converted to dex only when min_api is over 26(Android Oreo).

* Proguard is disabled for test modules.

Bug: 69899800
Test: mma -j under packages/apps/TV

Change-Id: Ic725f9b551ac9d5c26bcfa879033a3581ea579ae