Remove unnecessary TimeUnit conversion (converting from one unit to the same unit).
This change is a no-op, but please make sure that this did not expose a hidden bug! Are you sure the original code intended for this statement to be a no op??

If you wish to further improve this code (in a follow up CL), consider using a java.time.Duration instead of a bare primitive. See http://go/java-practices/time#avoid-primitives

Tested:
    TAP --sample for global presubmit queue
    http://test/OCL:185213730:BASE:185210942:1518228182204:44ec5277
PiperOrigin-RevId: 185369428
Change-Id: Ifaf7d9e8f75a5f1631ecb650654414ffe556c739